swich to gcc 5.3 and copypaste mpreal.h from path to sources

+++ src/lib/mesh/DataBus.hpp
@@ -50,34 +50,34 @@ template<typename Smth>
 void DataBus<TModel, CubicGrid<Dimensionality>, TMaterial>::
 exchangeSomethingWithNeighbors(Mesh* mesh, std::vector<Smth>& vec) {
 
-	int bufferSize = mesh->borderSize * mesh->indexMaker(0);
+    size_t bufferSize = (size_t) (mesh->borderSize * mesh->indexMaker(0));
 	size_t size = vec.size();
 
 	if (Mpi::Rank() == 0) {
-		MPI_Sendrecv(&(vec[size - 2 * bufferSize]), (int) sizeof(Smth) * bufferSize,
+        MPI_Sendrecv(&(vec[size - 2 * bufferSize]), (int) (sizeof(Smth) * bufferSize),
 		             MPI_BYTE,
 		             Mpi::Rank() + 1, 1,
-		             &(vec[size - bufferSize]), (int) sizeof(Smth) * bufferSize, MPI_BYTE,
+                     &(vec[size - bufferSize]), (int) (sizeof(Smth) * bufferSize), MPI_BYTE,
 		             Mpi::Rank() + 1, 1,
 		             MPI::COMM_WORLD, MPI_STATUS_IGNORE);
 
 	} else if (Mpi::Rank() == Mpi::Size() - 1) {
-		MPI_Sendrecv(&(vec[(size_t)bufferSize]), (int) sizeof(Smth) * bufferSize, MPI_BYTE,
+        MPI_Sendrecv(&(vec[bufferSize]), (int) (sizeof(Smth) * bufferSize), MPI_BYTE,
 		             Mpi::Rank() - 1, 1,
-		             &(vec[0]), (int) sizeof(Smth) * bufferSize, MPI_BYTE,
+                     &(vec[0]), (int) (sizeof(Smth) * bufferSize), MPI_BYTE,
 		             Mpi::Rank() - 1, 1,
 		             MPI::COMM_WORLD, MPI_STATUS_IGNORE);
 
 	} else {
-		MPI_Sendrecv(&(vec[size - 2 * bufferSize]), (int) sizeof(Smth) * bufferSize,
+        MPI_Sendrecv(&(vec[size - 2 * bufferSize]), (int) (sizeof(Smth) * bufferSize),
 		             MPI_BYTE,
 		             Mpi::Rank() + 1, 1,
-		             &(vec[size - bufferSize]), (int) sizeof(Smth) * bufferSize, MPI_BYTE,
+                     &(vec[size - bufferSize]), (int) (sizeof(Smth) * bufferSize), MPI_BYTE,
 		             Mpi::Rank() + 1, 1,
 		             MPI::COMM_WORLD, MPI_STATUS_IGNORE);
-		MPI_Sendrecv(&(vec[(size_t)bufferSize]), (int) sizeof(Smth) * bufferSize, MPI_BYTE,
+        MPI_Sendrecv(&(vec[bufferSize]), (int) (sizeof(Smth) * bufferSize), MPI_BYTE,
 		             Mpi::Rank() - 1, 1,
-		             &(vec[0]), (int) sizeof(Smth) * bufferSize, MPI_BYTE,
+                     &(vec[0]), (int) (sizeof(Smth) * bufferSize), MPI_BYTE,
 		             Mpi::Rank() - 1, 1,
 		             MPI::COMM_WORLD, MPI_STATUS_IGNORE);
 	}
